+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+-   Fix Hades window being positioned offscreen when switching to windowed mode while using a custom resolution larger than officially supported by the main monitor (multi-monitor use case).
 -   Fix load screen transition clipping when passing through door Styx -> [Redacted] for 32-bit engine.
 -   Fix Hades directory auto-detection not expanding user variables on MacOS / Linux.
 -   Fix Hades directory auto-detection not properly picking up Steam libraries.